+++ Engine/DogBackward.h
@@ -8,12 +8,12 @@ public:
 	void ClamptoScreen();
 	void Draw(Graphics& gfx) const;
 	void Update(const Keyboard& kbd);
-	int GetX() const;
-	int GetY() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetHeight() const;
 private:
-	int x = 399;
-	int y = 299;
-	int Width = 16;
-	int Height = 15;
+	float x = 399.0f;
+	float y = 299.0f;
+	float Width = 16.0f;
+	float Height = 15.0f;
 };
